[refactor] user 4계층 구조화
