+++ user/src/main/java/com/sparta/rooibos/user/domain/critria/UserSearchCriteria.java
@@ -1,0 +1,8 @@
+package com.sparta.rooibos.user.domain.critria;
+
+public record UserSearchCriteria(String keyword,
+                                 String sort,
+                                 String filter,
+                                 int page,
+                                 int size) {
+}
