+++ user/src/main/java/com/sparta/rooibos/user/domain/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package com.sparta.rooibos.user.domain.repository;
 
 import com.sparta.rooibos.user.application.dto.request.UserSearchRequest;
+import com.sparta.rooibos.user.domain.critria.UserSearchCriteria;
 import com.sparta.rooibos.user.domain.entity.User;
 import com.sparta.rooibos.user.domain.model.Pagination;
 
@@ -18,5 +19,5 @@ public interface UserRepository {
 
     void deleteById(UUID id);
 
-    Pagination<User> searchUsers(UserSearchRequest request);
+    Pagination<User> searchUsers(UserSearchCriteria criteria);
 }
