+++ user/src/main/java/com/sparta/rooibos/user/infrastructure/persistence/UserRepositoryCustomImpl.java
@@ -5,6 +5,7 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.sparta.rooibos.user.application.dto.request.UserSearchRequest;
+import com.sparta.rooibos.user.domain.critria.UserSearchCriteria;
 import com.sparta.rooibos.user.domain.entity.QUser;
 import com.sparta.rooibos.user.domain.entity.Role;
 import com.sparta.rooibos.user.domain.entity.User;
@@ -22,19 +23,19 @@ public class UserRepositoryCustomImpl implements UserRepositoryCustom {
     private final JPAQueryFactory queryFactory;
 
     @Override
-    public Pagination<User> searchUsers(UserSearchRequest request) {
+    public Pagination<User> searchUsers(UserSearchCriteria criteria) {
         QUser user = QUser.user;
         BooleanBuilder whereClause = new BooleanBuilder();
 
         // 필터별 키워드 검색
-        if (request.keyword() != null && !request.keyword().isBlank()) {
-            switch (request.filter()) {
-                case "username" -> whereClause.and(user.username.containsIgnoreCase(request.keyword()));
-                case "email" -> whereClause.and(user.email.containsIgnoreCase(request.keyword()));
-                case "slackAccount" -> whereClause.and(user.slackAccount.containsIgnoreCase(request.keyword()));
+        if (criteria.keyword() != null && !criteria.keyword().isBlank()) {
+            switch (criteria.filter()) {
+                case "username" -> whereClause.and(user.username.containsIgnoreCase(criteria.keyword()));
+                case "email" -> whereClause.and(user.email.containsIgnoreCase(criteria.keyword()));
+                case "slackAccount" -> whereClause.and(user.slackAccount.containsIgnoreCase(criteria.keyword()));
                 case "role" -> {
                     try {
-                        Role role = Role.valueOf(request.keyword().toUpperCase());
+                        Role role = Role.valueOf(criteria.keyword().toUpperCase());
                         whereClause.and(user.role.eq(role));
                     } catch (IllegalArgumentException ignored) {
                         whereClause.and(user.role.isNull());
@@ -44,7 +45,7 @@ public class UserRepositoryCustomImpl implements UserRepositoryCustom {
         }
 
         // 정렬 설정 (기본값: desc)
-        OrderSpecifier<?> orderSpecifier = "asc".equalsIgnoreCase(request.sort())
+        OrderSpecifier<?> orderSpecifier = "asc".equalsIgnoreCase(criteria.sort())
                 ? user.createdAt.asc()
                 : user.createdAt.desc();
 
@@ -52,8 +53,8 @@ public class UserRepositoryCustomImpl implements UserRepositoryCustom {
                 .selectFrom(user)
                 .where(whereClause)
                 .orderBy(orderSpecifier)
-                .offset((long) (request.page() - 1) * request.size())
-                .limit(request.size())
+                .offset((long) (criteria.page() - 1) * criteria.size())
+                .limit(criteria.size())
                 .fetch();
 
         Long total = queryFactory
@@ -62,6 +63,6 @@ public class UserRepositoryCustomImpl implements UserRepositoryCustom {
                 .where(whereClause)
                 .fetchOne();
 
-        return Pagination.of(request.page(), request.size(), total != null ? total : 0, results);
+        return Pagination.of(criteria.page(), criteria.size(), total != null ? total : 0, results);
     }
 }
